+++ release/src/index.ts
@@ -1,7 +1,7 @@
 import { EthereumWalletMock, ethereumWalletMockFixtures } from '@synthetixio/ethereum-wallet-mock'
 import { defineWalletSetup } from '../../packages/cache'
 import { MetaMask, getExtensionId, metaMaskFixtures } from '@synthetixio/synpress-metamask'
-import { testWithSynpress } from '@synthetixio/synpress-test'
+import { testWithSynpress } from '@synthetixio/synpress-core'
 
 export {
   // Framework fixtures