+++ wallets/ethereum-wallet-mock/test/synpress.ts
@@ -1,4 +1,4 @@
-import { testWithSynpress } from '@synthetixio/synpress-test'
+import { testWithSynpress } from '@synthetixio/synpress-core'
 import { ethereumWalletMockFixtures } from '../src'
 
 export default testWithSynpress(ethereumWalletMockFixtures)