+++ wallets/metamask/test/synpress.ts
@@ -1,4 +1,4 @@
-import { testWithSynpress } from '@synthetixio/synpress-test'
+import { testWithSynpress } from '@synthetixio/synpress-core'
 import { metaMaskFixtures } from '../src'
 import connectedSetup from './wallet-setup/connected.setup'