+++ packages/core/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @synthetixio/synpress-test
+# @synthetixio/synpress-core
 
 ## 0.0.1-alpha.3