+++ wallets/metamask/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Alpha Release
 - Updated dependencies
   - @synthetixio/synpress-cache@0.0.1-alpha.3
-  - @synthetixio/synpress-test@0.0.1-alpha.3
+  - @synthetixio/synpress-core@0.0.1-alpha.3
 
 ## 0.0.1-alpha.2
 
@@ -16,7 +16,7 @@
 - Alpha Release
 - Updated dependencies
   - @synthetixio/synpress-cache@0.0.1-alpha.2
-  - @synthetixio/synpress-test@0.0.1-alpha.2
+  - @synthetixio/synpress-core@0.0.1-alpha.2
 
 ## 0.0.1-alpha.1
 
@@ -25,7 +25,7 @@
 - Alpha Release
 - Updated dependencies
   - @synthetixio/synpress-cache@0.0.1-alpha.1
-  - @synthetixio/synpress-test@0.0.1-alpha.1
+  - @synthetixio/synpress-core@0.0.1-alpha.1
 
 ## 0.0.1-alpha.0